requirements.txt
google-generativeai
PyMuPDF
streamlit
python-dotenv